Adds support for `webpackConfig` option being an object.
Adds a check to see if `webpackConfig` is a string, and resolves it if so; otherwise the provided object is provided to `resolver.resolve` as-is.
Creates the setting for `resolver.resolve` once, in the constructor, and re-uses them with each invocation of `resolveFile`.
Adds tests for the case where `webpackConfig` is an object.  The snapshot has been confirmed to be identical to the case where `webpackConfig` is a string.