Fixes an incorrect transform for confusingly named exports.
Previously, as an optimization step, I assume, an import with the same name was searched for and, if found, the search for exports would be skipped and the import would be used to continue the search.
This runs on the assumption that a programmer would never create an export with a name that conflicts with an imported name.  This is probably true in almost every case, but if it isn't, the program is broken.  It is a dangerous assumption.
This change fixes this issue and creates a test for verification.