Prevents re-visitation within the same program.
A Babel plugin's traverser always visits every node, including nodes that have been replaced or added from a previous transformation.
If no state is used to keep track of what visits are necessary, an `ImportDeclaration` freshly created by this program will be visited.  This causes an unnecessary recursion and costs time, re-processing nodes that we know are already correct.
It also creates the potential for infinite recursions, where an import is visited, replaced, then the new node is visited again; repeat this for ever.  I've run into this issue many times as I tried to write new code for this plugin.
This adds a `Program` visitor, which initializes a set of `visitedNames` which is populated as the `ImportDeclaration` visitor performs its work.  If a node is recognized within this set, it will be skipped.