Fixes a heisenbug with `export-from` syntax.
When an export specifier with a path was found, it was incorrectly using `importName` to set the next name to search for, but an export specifier has no `importName` property.
I had noticed this typo quite some time ago, but for the life of me, I could not craft a test that would fail.  As is usual, it was a combination of problems that were acting to cover the bug up and continue to produce correct output.
First was the fact that the `ImportDeclaration` visitor would visit freshly inserted import declarations; this meant that when the visitor would fail to completely resolve a module, it would stop and use last good specifier it had to re-write the import, then it would get another chance to resolve it with fresh state when the re-written import declaration was visited.  Eventually, it would visit this import enough times to fully resolve it.
But, with the addition of the `visitedNames` cache, this bug should have been found, right?  Well, the issue here was the tests.  The `testmodule` always gave the final module a new alias each time it was re-exported, so it would always find a new name to search for.
So, the test that was added purposely always re-exports the module without altering its name.
Test created, test failed, bug fixed, test passed!