Caches specifiers from modules.
The only real purpose of the `AST` class is to act as friendly access to a file's import/export specifiers, and each time they are accessed, an AST is searched.
This change deprecates the `AST` class in favor of the `SpecResolver`, which will parse a file's AST, derive its import/export specifiers, then cache the results for later lookup.
For better consistency, a `PathResolver` class was also added.  The old `Resolver` class still exists, but is just a simple shim around `PathResolver`.
Both `AST` and `Resolver` are marked as deprecated.