Refactor to `specResolver` and the specifier extractors.
Simplifies `SpecResolver#resolve`, separating the caching logic from the logic that finds the specifiers.  The latter logic is now in `SpecResolver#getSpecifiers`.
Types are now checked using `babel-types` instead of a string comparison on the node's `type` property.  This should be a little more future proof, in case the way the AST is structured changes in the future.
`extractExportSpecifiers` now correctly handles the `ExportDefaultDeclaration` node type, which has no `specifiers` property and must be transformed differently.
The imported/exported name of a default import/export is now `default`.