Makes transforming `default` imports an option.
Provides a new `transformDefaultImports` option that allows a user to configure the plugin to follow default imports and exports.  The default for the option is `false`, which retains the original behavior.
These rules now apply to all `default` imports and exports, and not just the exports of the file being transformed by the plugin.
Adds tests for the new behavior.
Updates the `README.md` file with a list of supported options.