Separates core plugin logic into a `core` module.
Adds extensive documentation to describe the changes.
Simplifies the `ImportDeclaration` visitor to a couple of `Array#map` calls.
Can now handle cases where parsing fails; when this happens, it is most likely due to a Webpack loader not being present to provide the intended transformation.
Adds a test to ensure that parsing failures are handled correctly.