Solves module concatenation de-optimization.
Now employs a second, internal `spec-loader` for getting the specifiers of a module.  This prevents the `ModuleConcatenationPlugin` from de-optimizing due to loading other modules via `loaderContext.loadModule()`.
Many formerly bare-object types, like `ExportSpecifier` and `ResolvedPath` are now proper classes.  These classes are designed with serialization in mind.
The way side-effects are handled is now different.  Side-effecting modules can now be transformed, however doing so may create a side-effecting import if the side-effecting module is no longer directly imported.  The `ignoreSideEffects` option was changed to `transformSideEffects` to reflect this change.
The `export-from` syntax is now transformed in the same way as imports.
Adds more error classes for detecting and responding to specific failure states.